@@ -1,4 +1,4 @@
-module.exports = {
+export default {
 	root: true,
 	parser: '@typescript-eslint/parser',
 	plugins: ['simple-import-sort'],
